backend/requirements.txt
alembic
psycopg2
fastapi
pydantic
uvicorn
sqlmodel
google-api-python-client
google-auth-httplib2
google-auth-oauthlib